% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{OpenRepGridGui}
\alias{OpenRepGridGui}
\alias{OpenRepGridGui-package}
\title{\pkg{OpenRepGridGui}: A collection of GUIs for OpenRepGrid}
\description{
\pkg{OpenRepGridGui}: A collection of GUIs for OpenRepGrid
}
\note{
Nothing
}
\author{
Mark Heckmann
}
\references{
None
}
\keyword{package}
\keyword{repgrid}

